refactor: rearrange controllers
