+++ internal/controller/meta/admin.go
@@ -1,4 +1,4 @@
-package controller
+package meta
 
 import (
 	"time"
