+++ internal/controller/meta/meta.go
@@ -1,4 +1,4 @@
-package controller
+package meta
 
 import (
 	"time"
